Django API/requirements.txt
django>=4.2.5
psycopg2-binary>=2.9.9
djangorestframework>=3.14.0
django-filter>=24.2
djangorestframework-simplejwt>=5.3.1
drf-nested-routers>=0.94.1